do another refactoring

+++ model/src/main/java/com/alevel/deliverit/customers/SenderId.java
@@ -2,8 +2,11 @@ package com.alevel.deliverit.customers;
 
 import com.alevel.deliverit.entity.EntityId;
 
-public class SenderId extends EntityId<Integer> {
-    protected SenderId(Integer value) {
+/**
+ *
+ */
+public class SenderId extends EntityId<Long> {
+    protected SenderId(Long value) {
         super(value);
     }
 }
